Added slate-edit-table
* added @joefitter/slate-edit-table forked from Gitbook.IO and patched to work with state@0.47
* refactored editor into plugins
