+++ client/components/editor/initial-value.json
@@ -1,0 +1,16 @@
+{
+  "document": {
+    "nodes": [
+      {
+        "object": "block",
+        "type": "paragraph",
+        "nodes": [
+          {
+            "object": "text",
+            "text": ""
+          }
+        ]
+      }
+    ]
+  }
+}
