+++ client/components/editor/marks/index.js
@@ -1,0 +1,9 @@
+import renderMark from './render-mark';
+import { onKeyDown } from './handlers';
+
+export default function Marks() {
+  return {
+    renderMark,
+    onKeyDown
+  }
+}
