+++ package.json
@@ -36,6 +36,7 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "@joefitter/docx": "^4.7.0",
+    "@joefitter/slate-edit-table": "0.0.1",
     "@ukhomeoffice/frontend-toolkit": "^2.8.1",
     "@ukhomeoffice/react-components": "^0.7.5",
     "babel-loader": "^8.0.5",
@@ -67,8 +68,11 @@
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.2",
-    "slate": "^0.44.9",
-    "slate-react": "^0.21.16",
+    "slate": "^0.47.4",
+    "slate-edit-list": "^0.12.1",
+    "slate-edit-table": "^0.18.0",
+    "slate-hyperscript": "^0.13.4",
+    "slate-react": "^0.22.4",
     "webpack": "^4.23.1",
     "webpack-cli": "^3.1.2"
   },
